refactor: 정책 클래스를 policy 패키지로 이동
